refactor endpoint logic
